Deal with haddock comments with GHC option -haddock
The haddock comments appear mostly twice, once as "normal" comments,
except structured as haddock comments, and the valid ones are inserted
into the AST using HsDocTy.

Because we do not have a guarantee that all the comments get
allocated, we ignore the in-tree allocated haddock comments and
work only with the normal ones.

Closes #125